requirements.txt
fuzzywuzzy==0.18.0
scipy==1.4.1
Flask==1.1.2
flask-restful==0.3.6
python-Levenshtein==0.12.0
